bring up to date with master

+++ surveyKPI/src/surveyKPI/UserList.java
@@ -41,11 +41,9 @@ import org.smap.sdal.managers.LogManager;
 import org.smap.sdal.managers.MessagingManager;
 import org.smap.sdal.managers.SurveyTableManager;
 import org.smap.sdal.managers.UserManager;
-import org.smap.sdal.model.Action;
-import org.smap.sdal.model.Project;
-import org.smap.sdal.model.Role;
 import org.smap.sdal.model.User;
 import org.smap.sdal.model.UserGroup;
+import org.smap.sdal.model.UserSimple;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -113,115 +111,20 @@ public class UserList extends Application {
 		a.isAuthorised(sd, request.getRemoteUser());
 		// End Authorisation
 		
-		String sql = "select id,"
-				+ "ident, "
-				+ "name, "
-				+ "email "
-				+ "from users "
-				+ "where users.o_id = ? "
-				+ "and not users.temporary "
-				+ "order by ident asc";
-		PreparedStatement pstmt = null;
-		
-		String sqlGroups = "select g.id,"
-				+ "g.name "
-				+ "from groups g,"
-				+ "user_group ug "
-				+ "where ug.u_id = ? "
-				+ "and ug.g_id = g.id "
-				+ "order by g.id asc";
-		PreparedStatement pstmtGroups = null;
-		
-		String sqlProjects = "select p.id,"
-				+ "p.name "
-				+ "from project p,"
-				+ "user_project up "
-				+ "where up.u_id = ? "
-				+ "and up.p_id = p.id "
-				+ "order by p.name asc";
-		PreparedStatement pstmtProjects = null;
-		
-		String sqlRoles = "select r.id,"
-				+ "r.name "
-				+ "from role r,"
-				+ "user_role ur "
-				+ "where ur.u_id = ? "
-				+ "and ur.r_id = r.id "
-				+ "order by r.name asc";
-		PreparedStatement pstmtRoles = null;
-				
-		ArrayList<User> users = new ArrayList<User> ();
+		ArrayList<User> users = null;
+		Gson gson = new GsonBuilder().disableHtmlEscaping().setDateFormat("yyyy-MM-dd HH:mm:ss").create();
 		
 		try {
-			int o_id = GeneralUtilityMethods.getOrganisationId(sd, request.getRemoteUser(), 0);
+			// Get the users locale
+			Locale locale = new Locale(GeneralUtilityMethods.getUserLanguage(sd, request, request.getRemoteUser()));
+			ResourceBundle localisation = ResourceBundle.getBundle("org.smap.sdal.resources.SmapResources", locale);	
+			
+			int oId = GeneralUtilityMethods.getOrganisationId(sd, request.getRemoteUser(), 0);
 			boolean isOrgUser = GeneralUtilityMethods.isOrgUser(sd, request.getRemoteUser());
 			boolean isSecurityManager = GeneralUtilityMethods.hasSecurityRole(sd, request.getRemoteUser());
 			
-			pstmt = sd.prepareStatement(sql);
-			ResultSet rs = null;
-
-			pstmtGroups = sd.prepareStatement(sqlGroups);
-			ResultSet rsGroups = null;
-			
-			pstmtProjects = sd.prepareStatement(sqlProjects);
-			ResultSet rsProjects = null;
-			
-			pstmtRoles = sd.prepareStatement(sqlRoles);
-			ResultSet rsRoles = null;
-			
-			pstmt.setInt(1, o_id);
-			log.info("Get user list: " + pstmt.toString());
-			rs = pstmt.executeQuery();
-			while(rs.next()) {
-				User user = new User();
-				
-				user.id = rs.getInt("id");
-				user.ident = rs.getString("ident");
-				user.name = rs.getString("name");
-				user.email = rs.getString("email");
-				
-				// Groups
-				if(rsGroups != null) try {rsGroups.close();} catch(Exception e) {};
-				pstmtGroups.setInt(1, user.id);
-				rsGroups = pstmtGroups.executeQuery();
-				user.groups = new ArrayList<UserGroup> ();
-				while(rsGroups.next()) {
-					UserGroup ug = new UserGroup();
-					ug.id = rsGroups.getInt("id");
-					ug.name = rsGroups.getString("name");
-					user.groups.add(ug);
-				}
-				
-				// Projects
-				if(rsProjects != null) try {rsProjects.close();} catch(Exception e) {};
-				pstmtProjects.setInt(1, user.id);
-				rsProjects = pstmtProjects.executeQuery();
-				user.projects = new ArrayList<Project> ();
-				while(rsProjects.next()) {
-					Project p = new Project();
-					p.id = rsProjects.getInt("id");
-					p.name = rsProjects.getString("name");
-					user.projects.add(p);
-				}
-				
-				// Roles
-				if(isOrgUser || isSecurityManager) {
-					if(rsRoles != null) try {rsRoles.close();} catch(Exception e) {};
-					pstmtRoles.setInt(1, user.id);
-					rsRoles = pstmtRoles.executeQuery();
-					user.roles = new ArrayList<Role> ();
-					while(rsRoles.next()) {
-						Role r = new Role();
-						r.id = rsRoles.getInt("id");
-						r.name = rsRoles.getString("name");
-						user.roles.add(r);
-					}
-				}
-				
-				users.add(user);
-			}
-			
-			Gson gson=  new GsonBuilder().disableHtmlEscaping().setDateFormat("yyyy-MM-dd HH:mm:ss").create();
+			UserManager um = new UserManager(localisation);
+			users = um.getUserList(sd, oId, isOrgUser, isSecurityManager);
 			String resp = gson.toJson(users);
 			response = Response.ok(resp).build();
 						
@@ -232,15 +135,57 @@ public class UserList extends Application {
 			response = Response.serverError().entity(e.getMessage()).build();
 		    
 		} finally {
-			try {if (pstmt != null) {pstmt.close();	}} catch (SQLException e) {	}
-			try {if (pstmtGroups != null) {pstmtGroups.close();	}} catch (SQLException e) {	}
-			try {if (pstmtProjects != null) {pstmtProjects.close();	}} catch (SQLException e) {	}
-			try {if (pstmtRoles != null) {pstmtRoles.close();	}} catch (SQLException e) {	}
+			
 			SDDataSource.closeConnection(requestName, sd);
 		}
 
 		return response;
 	}
+	
+	@GET
+	@Produces("application/json")
+	@Path("/simple")
+	public Response getUsersSimple(
+			@Context HttpServletRequest request
+			) { 
+
+		Response response = null;
+		String requestName = "surveyKPI-getUsers";
+		
+		// Authorisation - Access
+		Connection sd = SDDataSource.getConnection(requestName);
+		a.isAuthorised(sd, request.getRemoteUser());
+		// End Authorisation
+		
+		ArrayList<UserSimple> users = null;
+		Gson gson = new GsonBuilder().disableHtmlEscaping().setDateFormat("yyyy-MM-dd HH:mm:ss").create();
+		
+		try {
+			// Get the users locale
+			Locale locale = new Locale(GeneralUtilityMethods.getUserLanguage(sd, request, request.getRemoteUser()));
+			ResourceBundle localisation = ResourceBundle.getBundle("org.smap.sdal.resources.SmapResources", locale);	
+			
+			int oId = GeneralUtilityMethods.getOrganisationId(sd, request.getRemoteUser(), 0);
+			
+			UserManager um = new UserManager(localisation);
+			users = um.getUserListSimple(sd, oId);
+			String resp = gson.toJson(users);
+			response = Response.ok(resp).build();
+						
+			
+		} catch (Exception e) {
+			
+			log.log(Level.SEVERE,"Error: ", e);
+			response = Response.serverError().entity(e.getMessage()).build();
+		    
+		} finally {
+			
+			SDDataSource.closeConnection(requestName, sd);
+		}
+
+		return response;
+	}
+
 
 	@GET
 	@Path("/temporary")
@@ -257,11 +202,17 @@ public class UserList extends Application {
 		// Authorisation - Access
 		Connection sd = SDDataSource.getConnection(requestName);
 		a.isAuthorised(sd, request.getRemoteUser());
-		// End Authorisation
-			
-		ActionManager am = new ActionManager();		
+		// End Authorisation			
 		
 		try {
+			// Localisation			
+			Locale locale = new Locale(GeneralUtilityMethods.getUserLanguage(sd, request, request.getRemoteUser()));
+			ResourceBundle localisation = ResourceBundle.getBundle("org.smap.sdal.resources.SmapResources", locale);
+			
+			String tz = "UTC";	// Set default for timezone
+			
+			ActionManager am = new ActionManager(localisation, tz);
+			
 			int o_id = GeneralUtilityMethods.getOrganisationId(sd, request.getRemoteUser(), 0);
 			Gson gson = new GsonBuilder().disableHtmlEscaping().setDateFormat("yyyy-MM-dd HH:mm:ss").create();
 			
@@ -400,11 +351,12 @@ public class UserList extends Application {
 		PreparedStatement pstmt = null;
 		try {	
 			
-			ResourceBundle localisation = null;
+			Locale locale = new Locale(GeneralUtilityMethods.getUserLanguage(sd, request, request.getRemoteUser()));
+			ResourceBundle localisation = ResourceBundle.getBundle("org.smap.sdal.resources.SmapResources", locale);
+			
 			String sql = null;
 			int o_id;
 			String adminName = null;
-			String language;	// Language spoken by user
 			ResultSet resultSet = null;
 			boolean isOrgUser = GeneralUtilityMethods.isOrgUser(sd, request.getRemoteUser());
 			boolean isSecurityManager = GeneralUtilityMethods.hasSecurityRole(sd, request.getRemoteUser());
@@ -412,7 +364,7 @@ public class UserList extends Application {
 			/*
 			 * Get the organisation and name of the user making the request
 			 */
-			sql = "SELECT u.o_id, u.name, u.language " +
+			sql = "SELECT u.o_id, u.name " +
 					" FROM users u " +  
 					" WHERE u.ident = ?;";				
 						
@@ -422,14 +374,7 @@ public class UserList extends Application {
 			resultSet = pstmt.executeQuery();
 			if(resultSet.next()) {
 				o_id = resultSet.getInt(1);
-				adminName = resultSet.getString(2);
-				language = resultSet.getString(3);
-				
-				// Set locale
-				if(language == null) {
-					language = "en";
-				}
-				localisation = ResourceBundle.getBundle("org.smap.sdal.resources.SmapResources", new Locale(language));
+				adminName = resultSet.getString(2);			
 				
 				for(int i = 0; i < uArray.size(); i++) {
 					User u = uArray.get(i);
@@ -439,7 +384,7 @@ public class UserList extends Application {
 						u.email = null;
 					}
 					
-					UserManager um = new UserManager();
+					UserManager um = new UserManager(localisation);
 					if(u.id == -1) {
 						// New user
 						um.createUser(sd, u, o_id,
@@ -452,7 +397,7 @@ public class UserList extends Application {
 								localisation);
 	
 						lm.writeLogOrganisation(sd, 
-								o_id, request.getRemoteUser(), "create", "User " + u.ident);
+								o_id, request.getRemoteUser(), "create", "User " + u.ident + " was created");
 								
 					} else {
 						// Existing user
@@ -461,10 +406,11 @@ public class UserList extends Application {
 								isSecurityManager,
 								request.getRemoteUser(),
 								request.getServerName(),
-								adminName);
+								adminName,
+								false);
 						
 						lm.writeLogOrganisation(sd, 
-								o_id, request.getRemoteUser(), "Update", "User " + u.ident);
+								o_id, request.getRemoteUser(), "Update", "User " + u.ident + " was updated. Groups: " + getGroups(u.groups));
 					}
 					
 					// Record the user change so that devices can be notified
@@ -525,6 +471,8 @@ public class UserList extends Application {
 		PreparedStatement pstmt = null;
 		PreparedStatement pstmtUpdate = null;
 		PreparedStatement pstmtGetIdent = null;
+		PreparedStatement pstmtCountOrgs = null;
+		PreparedStatement pstmtSoftDelete = null;
 		String basePath = GeneralUtilityMethods.getBasePath(request);
 		
 		try {	
@@ -548,11 +496,21 @@ public class UserList extends Application {
 					+ "and u.o_id = ?";								
 			pstmtGetIdent = sd.prepareStatement(sqlGetIdent);
 			
-			// Delete the user
-			String sqlUpdate = "DELETE FROM users u " +  
-					" WHERE u.id = ? " +			// Ensure the user is in the same organisation as the administrator doing the editing
-					" AND u.o_id = ?;";					
-			pstmtUpdate = sd.prepareStatement(sqlUpdate);
+			// Get the count of organisations that the user is in
+			String sqlCountOrgs = "select count(*) from user_organisation where u_id = ?";
+			pstmtCountOrgs = sd.prepareStatement(sqlCountOrgs);
+			
+			// Perform a hard delete of the user
+			String sqlUpdate = "delete from users u "  
+					+ "where u.id = ? "		// Ensure the user is in the same organisation as the administrator doing the editing
+					+ "and u.o_id = ?";					
+			pstmtUpdate = sd.prepareStatement(sqlUpdate);	
+			
+			// Perform a soft delete of the user
+			String sqlSoftDelete = "delete from user_organisation "  
+					+ "where u_id = ? "		// Ensure the user is in the same organisation as the administrator doing the editing
+					+ "and o_id = ?";					
+			pstmtSoftDelete = sd.prepareStatement(sqlSoftDelete);	
 			
 			// Get the organisation id
 			pstmt.setString(1, request.getRemoteUser());
@@ -573,21 +531,44 @@ public class UserList extends Application {
 						ident = rs.getString(1);
 					}
 					
-					// Peform the delete
-					pstmtUpdate.setInt(1, u.id);
-					pstmtUpdate.setInt(2, o_id);
-					log.info("Delete user: " + pstmt.toString());
+					// Get the number of organisations that this user is a member of
+					int numberOrgs = 0;
+					pstmtCountOrgs.setInt(1, u.id);
+					rs = pstmtCountOrgs.executeQuery();
+					if(rs.next()) {
+						numberOrgs = rs.getInt(1);
+					}
 					
-					int count = pstmtUpdate.executeUpdate();
-					
-					if(count > 0) {	
-						// If a user was deleted then delete their directories
-						GeneralUtilityMethods.deleteDirectory(basePath + "/media/users/" + u.id);
+					if(numberOrgs <= 1) {
+						// Only one organisation so performa a Hard delete
+						pstmtUpdate.setInt(1, u.id);
+						pstmtUpdate.setInt(2, o_id);
+						log.info("Delete user: " + pstmtUpdate.toString());
 						
-						// Delete any csv table definitions that they have
-						SurveyTableManager stm = new SurveyTableManager(sd, localisation);
-						stm.deleteForUsers(ident);			// Delete references to this survey in the csv table 
-					}	
+						int count = pstmtUpdate.executeUpdate();
+						
+						if(count > 0) {	
+							// If a user was deleted then delete their directories
+							GeneralUtilityMethods.deleteDirectory(basePath + "/media/users/" + u.id);
+							
+							// Delete any csv table definitions that they have
+							SurveyTableManager stm = new SurveyTableManager(sd, localisation);
+							stm.deleteForUsers(ident);			// Delete references to this survey in the csv table 
+							
+							lm.writeLogOrganisation(sd, 
+									o_id, request.getRemoteUser(), "delete", "User " + u.ident + " was completely deleted " + o_id);
+						}	
+					} else {
+						// soft delete
+						pstmtSoftDelete.setInt(1, u.id);
+						pstmtSoftDelete.setInt(2, o_id);
+						log.info("Soft Delete user: " + pstmtSoftDelete.toString());
+						
+						 pstmtSoftDelete.executeUpdate();
+						 
+						lm.writeLogOrganisation(sd, 
+									o_id, request.getRemoteUser(), "delete", "User " + u.ident + " was soft deleted from organisation " + o_id);
+					}
 
 				}
 		
@@ -612,6 +593,8 @@ public class UserList extends Application {
 			try {if (pstmt != null) {pstmt.close();}} catch (SQLException e) {}
 			try {if (pstmtUpdate != null) {pstmtUpdate.close();}} catch (SQLException e) {}
 			try {if (pstmtGetIdent != null) {pstmtGetIdent.close();}} catch (SQLException e) {}
+			try {if (pstmtCountOrgs != null) {pstmtCountOrgs.close();}} catch (SQLException e) {}
+			try {if (pstmtSoftDelete != null) {pstmtSoftDelete.close();}} catch (SQLException e) {}
 			
 			SDDataSource.closeConnection(requestName, sd);
 		}
@@ -619,6 +602,38 @@ public class UserList extends Application {
 		return response;
 	}
 	
+	private String getGroups(ArrayList<UserGroup> groups) {
+		StringBuffer g = new StringBuffer("");
+		if(groups != null) {
+			for(UserGroup ug : groups) {
+				String name = null;
+				switch(ug.id) {
+					case 1: name = "admin";
+							break;
+					case 2: name = "analyst";
+							break;
+					case 3: name = "enum";
+							break;
+					case 4: name = "org admin";
+							break;
+					case 5: name = "manage";
+							break;
+					case 6: name = "security";
+							break;
+					case 7: name = "view data";
+							break;
+				}
+				if(name != null) {
+					if(g.length() > 0) {
+						g.append(", ");
+					}
+					g.append(name);
+				}
+				
+			}
+		}
+		return g.toString();
+	}
 
 }
